perf(home): cut homepage critical-path JS and image work
Defer announcement modal chunks, dynamic-import near-fold bands, move the hero LCP path off Framer Motion, and drop unused Swiper so first paint stays lighter.
