+++ docs/performance/lighthouse-issue-86-after.md
@@ -1,0 +1,12 @@
+# Lighthouse — issue-86-after
+
+Run at: 2026-07-24T01:33:58.754Z
+Base URL: http://127.0.0.1:3100
+
+| Route | Perf | A11y | BP | SEO | LCP | CLS | TBT |
+|---|---:|---:|---:|---:|---|---|---|
+| `/` | 42 | 96 | 100 | 92 | 4.6 s | 0.587 | 810 ms |
+| `/news` | 52 | 94 | 100 | 92 | 5.9 s | 0.587 | 110 ms |
+| `/about/team` | 40 | 96 | 100 | 92 | 4.2 s | 0.587 | 740 ms |
+| `/programs` | 54 | 96 | 100 | 92 | 4.8 s | 0.587 | 280 ms |
+| `/programs/akomapa-ghltp` | 49 | 90 | 100 | 92 | 4.8 s | 0.599 | 370 ms |
