+++ package.json
@@ -64,7 +64,6 @@
     "react-pdf": "^9.2.1",
     "sonner": "^2.0.3",
     "stripe": "^18.1.1",
-    "swiper": "12.2.0",
     "tailwind-merge": "^3.2.0",
     "tw-animate-css": "^1.2.5",
     "zod": "^3.24.2"
